Move mocks to the tests folder
