+++ listenbrainz/webserver/static/js/src/APIService.test.ts
@@ -1,7 +1,7 @@
 import APIService from "./APIService";
 
-const feedProps = require("./__mocks__/feedProps.json");
-const pinProps = require("./__mocks__/pinProps.json");
+const feedProps = require("../tests/__mocks__/feedProps.json");
+const pinProps = require("../tests/__mocks__/pinProps.json");
 
 const apiService = new APIService("foobar");
