+++ listenbrainz/webserver/static/js/src/playlists/Playlist.test.tsx
@@ -3,7 +3,7 @@ import { shallow, mount } from "enzyme";
 import * as timeago from "time-ago";
 import AsyncSelect from "react-select/async";
 import PlaylistPage from "./Playlist";
-import * as playlistPageProps from "../__mocks__/playlistPageProps.json";
+import * as playlistPageProps from "../../tests/__mocks__/playlistPageProps.json";
 import GlobalAppContext, { GlobalAppContextT } from "../GlobalAppContext";
 import APIService from "../APIService";
 import { MUSICBRAINZ_JSPF_PLAYLIST_EXTENSION } from "./utils";
