+++ listenbrainz/webserver/static/js/src/stats/UserArtistMap.test.tsx
@@ -3,9 +3,9 @@ import { mount, shallow } from "enzyme";
 
 import UserArtistMap, { UserArtistMapProps } from "./UserArtistMap";
 import APIError from "../APIError";
-import * as userArtistMapResponse from "../__mocks__/userArtistMap.json";
-import * as userArtistMapProcessedDataArtist from "../__mocks__/userArtistMapProcessDataArtist.json";
-import * as userArtistMapProcessedDataListen from "../__mocks__/userArtistMapProcessDataListen.json";
+import * as userArtistMapResponse from "../../tests/__mocks__/userArtistMap.json";
+import * as userArtistMapProcessedDataArtist from "../../tests/__mocks__/userArtistMapProcessDataArtist.json";
+import * as userArtistMapProcessedDataListen from "../../tests/__mocks__/userArtistMapProcessDataListen.json";
 
 const props: UserArtistMapProps = {
   user: {
