+++ listenbrainz/webserver/static/js/tests/__mocks__/colorPlayProps.json
