+++ listenbrainz/webserver/static/js/tests/__mocks__/getInfo.json
