+++ listenbrainz/webserver/static/js/tests/__mocks__/lookupMBRelease.json
