+++ listenbrainz/webserver/static/js/tests/__mocks__/userArtistMap.json
