+++ listenbrainz/webserver/static/js/tests/__mocks__/userArtistMapProcessDataArtist.json
