+++ listenbrainz/webserver/static/js/tests/__mocks__/userArtistMapProcessDataListen.json
