+++ listenbrainz/webserver/static/js/tests/__mocks__/userArtists.json
