+++ listenbrainz/webserver/static/js/tests/__mocks__/userDailyActivity.json
