+++ listenbrainz/webserver/static/js/tests/__mocks__/userDailyActivityProcessData.json
