+++ listenbrainz/webserver/static/js/tests/__mocks__/userFeedbackAPIResponse.json
