+++ listenbrainz/webserver/static/js/tests/__mocks__/userListeningActivityMonth.json
