+++ listenbrainz/webserver/static/js/tests/__mocks__/userListeningActivityProcessDataMonth.json
