+++ listenbrainz/webserver/static/js/tests/__mocks__/userListeningActivityProcessDataWeek.json
