+++ listenbrainz/webserver/static/js/tests/__mocks__/userListeningActivityProcessDataYear.json
