+++ listenbrainz/webserver/static/js/tests/__mocks__/userListeningActivityWeek.json
