+++ listenbrainz/webserver/static/js/tests/__mocks__/userListeningActivityYear.json
