+++ listenbrainz/webserver/static/js/tests/__mocks__/userReleases.json
